libraries/logging-settings/requirements.txt
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-json-logger==2.0.7
